+++ WSAdmin/DataModels/Invoice.swift
@@ -11,9 +11,9 @@ import Foundation
 //
 @Observable class Invoice: Identifiable {
 	var invoiceLines = [InvoiceLine]()
-	var totalCost: Float = 0.0
-	var totalRevenue: Float = 0.0
-	var totalProfit: Float = 0.0
+	var totalCost: Double = 0.0
+	var totalRevenue: Double = 0.0
+	var totalProfit: Double = 0.0
 	var totalSessions: Int = 0
 	var monthName: String = ""
 	var isInvoiceLoaded: Bool
