+++ WSAdmin/Views/ServiceView.swift
@@ -21,12 +21,12 @@ struct ServiceView: View {
 	@State var serviceType: ServiceTypeOption
 	@State var billingType: BillingTypeOption
 	@State var serviceCount: Int
-	@State var cost1: Float
-	@State var cost2: Float
-	@State var cost3: Float
-	@State var price1: Float
-	@State var price2: Float
-	@State var price3: Float
+	@State var cost1: Double
+	@State var cost2: Double
+	@State var cost3: Double
+	@State var price1: Double
+	@State var price2: Double
+	@State var price3: Double
 	
 	@State private var showAlert:Bool = false
